Add logger module.
